Add convenience wrapper around Registry.OnGlobal.
On the client object, so we don't need to do GetDisplay().GetRegistry()
ourselves, and it passes in an object, which we can then type-switch on
to get a proxy with all the methods we need.

TODO: give objects a way of getting the interface name/version.

This builds, but is untested; I need to rush out the door.